Separate consumers/producers/partitioners
