+++ kafka/consumer/__init__.py
@@ -1,0 +1,6 @@
+from .simple import SimpleConsumer
+from .multiprocess import MultiProcessConsumer
+
+__all__ = [
+    'SimpleConsumer', 'MultiProcessConsumer'
+]
