+++ kafka/partitioner/__init__.py
@@ -1,0 +1,6 @@
+from .roundrobin import RoundRobinPartitioner
+from .hashed import HashedPartitioner
+
+__all__ = [
+    'RoundRobinPartitioner', 'HashedPartitioner'
+]
