+++ kafka/producer/__init__.py
@@ -1,0 +1,6 @@
+from .simple import SimpleProducer
+from .keyed import KeyedProducer
+
+__all__ = [
+    'SimpleProducer', 'KeyedProducer'
+]
