+++ test/test_consumer_integration.py
@@ -4,7 +4,7 @@ from six.moves import xrange
 
 from kafka import SimpleConsumer, MultiProcessConsumer, create_message
 from kafka.common import ProduceRequest, ConsumerFetchSizeTooSmall
-from kafka.consumer import MAX_FETCH_BUFFER_SIZE_BYTES
+from kafka.consumer.base import MAX_FETCH_BUFFER_SIZE_BYTES
 
 from test.fixtures import ZookeeperFixture, KafkaFixture
 from test.testutil import (
